LogRetriever/src/test/java/com/sumeet/cribl/logretriever/tests/unit/LogServiceTest.java
package com.sumeet.cribl.logretriever.tests.unit;

class LogServiceTest {

    /*@Mock
    private LogStrategyFactory strategyFactory;
    @Mock
    private LogFilterFactory filterFactory;
    @Mock
    private LogRetrievalStrategy strategy;
    @Mock
    private LogFilter filter;

    private LogService logService;



    @BeforeEach
    void setUp() {
        MockitoAnnotations.openMocks(this);
        logService = new LogService(strategyFactory, filterFactory);
    }

    @Test
    void retrieveLogsAsync_Success() throws Exception {
        String filename = "test.log";
        String keyword = "error";
        Integer limit = 10;
        Path mockPath = mock(Path.class);
        List<List<String>> expectedLogs = Arrays.asList(
                Arrays.asList("log1", "log2"),
                Arrays.asList("log3", "log4")
        );

        when(strategyFactory.getStrategy(any())).thenReturn(strategy);
        when(filterFactory.createFilter(keyword, limit)).thenReturn(filter);
        when(strategy.retrieveLogsAsync(any(), any())).thenReturn(CompletableFuture.completedFuture(expectedLogs));

        CompletableFuture<List<List<String>>> result = logService.retrieveLogsAsync(filename, keyword, limit);

        assertNotNull(result);
        assertEquals(expectedLogs, result.get());
    }

    @Test
    void retrieveLogsAsync_FileNotFound() throws Exception {
        String filename = "nonexistent.log";
        when(logConfig.getLogDirectory()).thenReturn("/logs");
        when(strategyFactory.getStrategy(any())).thenThrow(new LogRetrievalException("File not found", new java.io.FileNotFoundException()));

        CompletableFuture<List<List<String>>> result = logService.retrieveLogsAsync(filename, null, null);

        assertNotNull(result);
        List<List<String>> logs = result.get();
        assertEquals(1, logs.size());
        assertEquals("Log file not found. Please check the filename and try again.", logs.get(0).get(0));
    }*/
}
